refactor: reuse auth forms
